Add files via upload
So I tried to make the potential energy positive, and it appears to have messed up the energy difference values. I would recommend just trying to track the potential and kinetic energies separately and see how they interact. So here's the RK4 code with potential energy values in the negative (because the 0 potential energy point is taken at the pivot). Please check for any mistake if you see any.